+++ packages/user/src/components/event/racing/Background.tsx
@@ -1,0 +1,9 @@
+export default function Background() {
+	return (
+		<img
+			className="absolute h-full w-full object-contain"
+			alt="레이싱 배경"
+			src="/images/racing/background.png"
+		/>
+	);
+}
