+++ packages/user/src/components/event/racing/RacingTitle.tsx
@@ -1,0 +1,15 @@
+export default function RacingTitle() {
+	return (
+		<>
+			<h3>
+				버튼을 연타해 승리를 <strong>CHARGE</strong>하세요!
+			</h3>
+			<div className="mb-5 flex items-center gap-3">
+				<p className="text-body-2 text-foreground/60">
+					1등에 가까워질 수 있도록 배터리를 가득 충전
+					<strong className="inline-block align-bottom leading-6">🔋</strong>해주세요!
+				</p>
+			</div>
+		</>
+	);
+}
