+++ packages/user/src/components/event/racing/teamGaugeButton/Gauge.tsx
