+++ packages/user/src/components/event/racing/teamGaugeButton/index.tsx
